# AtlasRoutes/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AtlasRoutes API",
        "description": "API for managing tourist itineraries, destinations, and user favorites.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "paths": {
...
        "/api/me/favorites": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "List current user favorites",
                "operationId": "12a0fb606b39cfe361ceac58162a79b5",
                "responses": {
                    "200": {
                        "description": "Favorites list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Favorite"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/itineraries/{itinerary}/favorite": {
            "post": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Add itinerary to favorites",
                "operationId": "63355e38e8c5643d880ebc2371c8c73f",
                "parameters": [
                    {
                        "name": "itinerary",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Favorite created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Added to favorites."
                                        },
                                        "favorite": {
                                            "$ref": "#/components/schemas/Favorite"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Already in favorites",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Already in favorites."
                                        },
                                        "favorite": {
                                            "$ref": "#/components/schemas/Favorite"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Remove itinerary from favorites",
                "operationId": "3634074ac69b65612fb937acd641b23a",
                "parameters": [
                    {
                        "name": "itinerary",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Favorite removed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Removed from favorites."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Activity": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "destination_id": {
                        "type": "integer",
                        "example": 4
                    },
                    "name": {
                        "type": "string",
                        "example": "Visit the old medina"
                    },
                    "type": {
                        "type": "string",
                        "example": "place",
                        "enum": [
                            "place",
                            "activity",
                            "dish"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2026-03-10T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2026-03-10T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "Destination": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 4
                    },
                    "itinerary_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "name": {
                        "type": "string",
                        "example": "Marrakesh"
                    },
                    "accommodation": {
                        "type": "string",
                        "example": "Riad Atlas",
                        "nullable": true
                    },
                    "activities": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Activity"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2026-03-10T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2026-03-10T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "Itinerary": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 2
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Atlas Route Highlights"
                    },
                    "category": {
                        "type": "string",
                        "example": "adventure"
                    },
                    "duration": {
                        "type": "integer",
                        "example": 5
                    },
                    "image": {
                        "type": "string",
                        "example": "https://example.com/atlas.jpg",
                        "nullable": true
                    },
                    "favorites_count": {
                        "type": "integer",
                        "example": 12
                    },
                    "destinations": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Destination"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2026-03-10T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2026-03-10T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "Favorite": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 7
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "itinerary_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "itinerary": {
                        "$ref": "#/components/schemas/Itinerary"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2026-03-10T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2026-03-10T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}